Create Python Realisation of QuickSort with Median of 3 pivot and Insertion Sort
This is a python realization of Quicksort, which chooses the pivot based on median of three. Also for array size lesser than 4 it uses insertion sort. That number 4 can be adjusted based on the size of the input array. This quicksort method is proposed by Sedgewick[1978].
This program asks user to input each element of the array. However, it can be modified for direct insertion of array.
Thanks for reading and suggestions are welsome.